+++ api/rs/slint/private_unstable_api.rs
@@ -235,6 +235,6 @@ pub mod re_exports {
     pub use unicode_segmentation::UnicodeSegmentation;
     pub use vtable::{self, *};
 
-    #[cfg(feature = "live-reload")]
-    pub use slint_interpreter::live_reload;
+    #[cfg(feature = "live-preview")]
+    pub use slint_interpreter::live_preview;
 }