+++ tests/driver/rust/build.rs
@@ -5,7 +5,7 @@ use std::io::{BufWriter, Write};
 use std::path::Path;
 
 fn main() -> std::io::Result<()> {
-    let live_reload = std::env::var("SLINT_LIVE_RELOAD").is_ok();
+    let live_preview = std::env::var("SLINT_LIVE_PREVIEW").is_ok();
 
     let mut generated_file = BufWriter::new(std::fs::File::create(
         Path::new(&std::env::var_os("OUT_DIR").unwrap()).join("generated.rs"),
@@ -21,17 +21,17 @@ fn main() -> std::io::Result<()> {
         let source = std::fs::read_to_string(&testcase.absolute_path)?;
         let ignored = if testcase.is_ignored("rust") {
             "#[ignore = \"testcase ignored for rust\"]"
-        } else if (cfg!(not(feature = "build-time")) || live_reload)
+        } else if (cfg!(not(feature = "build-time")) || live_preview)
             && source.contains("//bundle-translations")
         {
             "#[ignore = \"translation bundle not working with the macro\"]"
-        } else if live_reload && source.contains("ComponentContainer") {
+        } else if live_preview && source.contains("ComponentContainer") {
             "#[ignore = \"ComponentContainer doesn't work with the interpreter\"]"
-        } else if live_reload && source.contains("#3464") {
+        } else if live_preview && source.contains("#3464") {
             "#[ignore = \"issue #3464 not fixed with the interpreter\"]"
-        } else if live_reload && module_name.contains("widgets_menubar") {
+        } else if live_preview && module_name.contains("widgets_menubar") {
             "#[ignore = \"issue #8454 causes crashes\"]"
-        } else if live_reload && module_name.contains("write_to_model") {
+        } else if live_preview && module_name.contains("write_to_model") {
             "#[ignore = \"Interpreted model don't forward to underlying models for anonymous structs\"]"
         } else {
             ""
@@ -74,7 +74,7 @@ fn main() -> std::io::Result<()> {
 
     // By default resources are embedded. The WASM example builds provide test coverage for that. This switch
     // provides test coverage for the non-embedding case, compiling tests without embedding the images.
-    if !live_reload {
+    if !live_preview {
         println!("cargo:rustc-env=SLINT_EMBED_RESOURCES=false");
     }